Crate serde_json_core

serde-json for no_std programs

This version of serde-json is aimed at applications that run on resource constrained devices.

Status

This crate is currently being passively maintained by @japaric. I (@japaric) will be merging PRs (bug fixes and features) as they appear without any long-term / design consideration. If you would like to take over the steering wheel (i.e. take ownership of this crate) send me an e-mail (see my GitHub profile).

Current features

The error type is a simple C like enum (less overhead, smaller memory footprint)

(De)serialization doesn't require memory allocations

Deserialization of integers doesn't go through u64; instead the string is directly parsed into the requested integer type. This avoids pulling in KBs of compiler intrinsics when targeting a non 64-bit architecture.

Supports deserialization of:

bool

Integers

str (This is a zero copy operation.) (*)

Option

Arrays

Tuples

Structs

C like enums

Supports serialization (compact format only) of:

bool

Integers

str (**)

Option

Arrays

Tuples

Structs

C like enums

(*) Deserialization of strings ignores escaped sequences. Escaped sequences might be supported in the future using a different Serializer as this operation is not zero copy.

(**) Serialization of strings doesn't escape stuff. This simply has not been implemented yet.

Planned features

(De)serialization from / into IO objects once core::io::{Read,Write} becomes a thing.

Non-features

This is explicitly out of scope

Anything that involves dynamic memory allocation

Like the dynamic Value type

MSRV

This crate is guaranteed to compile on stable Rust 1.31.0 and up. It might compile with older versions but that may change in any new patch release.

Modules

de
Deserialize JSON data to a Rust data structure

ser
Serialize a Rust data structure into JSON data

Functions

from_slice
Deserializes an instance of type T from bytes of JSON text

from_str
Deserializes an instance of type T from a string of JSON text

to_string
Serializes the given data structure as a string of JSON text

to_vec
Serializes the given data structure as a JSON byte vector